+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 0.1.4
 - implement logout support
+- add CSRF protection for auth and logout endpoints
+- update `fkooman/rest`
 
 ## 0.1.3
 - support `application/x-www-form-urlencoded` response to verify request again